+++ CHANGELOG.md
@@ -20,10 +20,12 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to use `len(sized) > 0` style checks
 - Forbids to use repeatable conditions: `flag or flag`
 - Forbids to write conditions like `not some > 1`
+- Forbids to use complex unwrapping compares like `x == x > 0`
 
 ### Bugfixes
 
 - Fixes incorrect line number for `Z331`
+- Fixes that `Z311` was not raising for multiple `not in` cases
 
 ### Misc